create network
